Added add_employee.jsp,DepartmentService,employees.jsp,update_employee.jsp

+++ employee-management-system/src/main/java/com/ems/entity/Department.java
@@ -17,6 +17,6 @@ import lombok.NoArgsConstructor;
 public class Department {
 
 	@Id
-	private Integer deptId;
+	private Long deptId;
 	private String deptName;
 }
